+++ src/features/basic/correct-commands/correct-commands.ts
@@ -1,5 +1,4 @@
 import { changeInputValue } from '@src/util';
-import { correctExchangeOrderCommand } from './exchange-order-commands';
 import { correctMaterialCommand } from './material-commands';
 import { correctPlanetCommand } from './planet-commands';
 import { correctShipCommand } from './ship-commands';
@@ -8,7 +7,6 @@ import { correctXitWeb } from './xit-web';
 import { correctXitArgs } from '@src/infrastructure/prun-ui/buffers';
 
 const transformers = [
-  correctExchangeOrderCommand,
   correctMaterialCommand,
   correctPlanetCommand,
   correctShipCommand,
@@ -16,6 +14,10 @@ const transformers = [
   correctXitWeb,
   correctXitArgs,
 ];
+
+export function addCommandTransformer(transform: (parts: string[]) => void) {
+  transformers.push(transform);
+}
 
 async function onSelectorReady(selector: HTMLElement) {
   const input: HTMLInputElement = await $(selector, C.PanelSelector.input);