+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Added
 
-- `correct-commands`: Type an exchange shortcut and a material ticker (`a dw`) to open the CXPO place-order buffer for that material on that exchange
+- `shortcut-placeorder`: Type an exchange shortcut and a material ticker (`a dw`) to open the CXPO place-order buffer for that material on that exchange
 
 ## 1.1.1